requirements.txt
mmcv-full==1.4.0
cityscapesscripts
matplotlib
mmcls>=0.20.1
numpy
packaging
prettytable
codecov
flake8
interrogate
pytest
xdoctest>=0.10.0
yapf
torch-scatter

-e mmseg/ops/tome
-e mmseg/ops
